pyproject.toml
[project]
name = "wbb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "llm>=0.27.1",
    "lxml-html-clean>=0.4.2",
    "matplotlib>=3.10.3",
    "newspaper4k>=0.9.3.1",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "playwright>=1.52.0",
    "requests-cache>=1.2.1",
    "requests-html>=0.10.0",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "selenium>=4.34.2",
    "shot-scraper>=1.8",
    "sqlite-utils>=3.38",
    "tldextract>=5.3.0",
]
